Отделил SerializerConfigExtesions от SerializerConfig
