+++ ObjectPrinting/ObjectPrinting.csproj
@@ -53,6 +53,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Core\ISerializerConfig.cs" />
+    <Compile Include="Extensions\SerializerConfigExtensions.cs" />
     <Compile Include="ObjectPrinter.cs" />
     <Compile Include="ObjectPrinterConfigs\PrintingConfig.cs" />
     <Compile Include="ObjectPrinterConfigs\SerializerConfig.cs" />
